Creation of LICENSE file to add a license
By default, any creative work without an explicit license, including Github repositories like yours, have the most restrictive license of them all: exclusive "all rights reserved" copyright. I assume that since you have your (awesome) software in a Github repository, it is your intent for people to use it, fork it and improve it. Thus, I'd strongly advise to add an open source license to facilitate this -- otherwise, it would be illegal to use this code in any way. If you want to allow any use of it, I'd recommend a MIT License (which is the one I'm including in this pull-request). I am happy to help clarify doubts or uncertainties around all this, or help choose another license if you don't like these (there are other options in Github's https://choosealicense.com).